A few refactorings -- mostly renaming to more conventional naming -- and immproved code documentation.
