+++ dispatcher/dispatcher.go
@@ -180,7 +180,7 @@ func handleExecuteRequest(msg kernel.Message, goExec *goexec.State) error {
 		replyContent["status"] = "ok"
 		replyContent["user_expressions"] = make(map[string]string)
 	} else {
-		name, value, traceback := goexec.Unwrap(executionErr)
+		name, value, traceback := goexec.JupyterErrorSplit(executionErr)
 		replyContent["status"] = "error"
 		replyContent["ename"] = name
 		replyContent["evalue"] = value
