+++ goexec/elementtype_string.go
@@ -5,7 +5,7 @@ package goexec
 import "strconv"
 
 func _() {
-	// An "invalid array index" compiler error signifies that the constant values have changed.
+	// An "invalid array index" compiler err signifies that the constant values have changed.
 	// Re-run the stringer command to generate them again.
 	var x [1]struct{}
 	_ = x[Invalid-0]
